Changed to save the replay rather than the complete request. Fixed troops expended on the replay to be the troops used to match with client. Using War battle type but setting stars to 0 to prevent crash.

+++ src/main/java/swcnoops/server/model/BattleAttributes.java
@@ -3,7 +3,7 @@ package swcnoops.server.model;
 import java.util.List;
 
 public class BattleAttributes {
-    public int battleEndedAt;
+    public long battleEndedAt;
     public int damagePercentage;
     public List<String> deathLog;
     public String deviceInfo;
